importing, inexact search
Implement importing from the previous LibreOffice DB. Could still be
improved to move punctuation out of the awkward way I've had to do
it--but that will only be okay when/if I figure out how to sort it
without using those, which is why I did it that way in the first place.

Inexact LIKE search. Really awesome.