fix nearby() so it doesn't return spurious entries
Previously the BETWEEN query, since it was working on strings, would
do funny things if searched with single digits. This shouldn't happen
if we enforce leading zeroes, but just in case, it's a wise idea to
check to make sure the values returned are proper, since the user
could potentially enter other strange queries and get equally strange
results.

Searching for ranges is no longer supported. Also, entering a non-
string for pagenum (what used to work) will currently crash the
program. We need validation.