pad with leading zeroes when importing
I believe this fixes the nearby() issue now that we have leading zeroes
for both TB 3-digit numbers and CB 2-digit numbers. However, I have not
tested this and will not count it fixed yet. Also, work remains in that
we should remove those numbers when outputting them to the screen to be
nice and neat.

(Can we use a similar leading zero function to accept user dates more
cheerfully? I suppose we could also implement a datetime parser which
would like it better.)

Also, it doesn't pad ranges, which might be an occasional issue. Making
sure that works is probably an advanced feature.