start on implementing events
I wound up changing the database format to include an 'events' table.
This will allow searching for events in a reasonable fashion in the
future, and besides it's just cleaner than storing a bunch of discrete
pieces of data in a single table cell.

I still need to *remove* the old events field. (This will be the next
commit, as it involves changing a number of functions in backend to
reflect its nonexistence, and since I forgot earlier I'll have to
reinitialize the DB again.)

We can now see events that exist, but do not currently have a way to
*write* them. (Well, I guess you could go into backend.py and write the
necessary function calls manually at the bottom of the file. That's the
new paradigm for user interaction.) Nor can we edit or delete them.
Also, no other parts of the program (search, for instance) are
event-aware yet.