continue events
You can now add and delete events. I realized in their present form,
there's basically no point in editing an event, since it's basically
just as fast to delete and re-add it.

Little things yet:
- Add dates this notebook was open to the top of the screen.
- Actually put this option on the main menu.
- Possibly set order and repositioning?
- Line wrapping.

Also, the save/rollback thing is so handy/cool/performance-enhancing
that I intend to put it into the notebooks menu as well. It should not
be possible to quit the program without saving, so that shouldn't be an
issue (barring a power failure or something).