todo updates
I totally feel like working more on this, but it's midnight and I should
really go to bed because I want to try to go to Northfield tomorrow
morning and get a haircut, if there's space for me to make an
appointment somewhat late.

That really doesn't belong in a commit message, does it? Who cares!