fix infinite loop bug
Was relying on variables not being set (they were set to None as default
arguments) for an if statement determining whether to get user input. If
the user entered invalid input, those variables would never be set, and
thus it would fall through to the validation condition evaluating false
again.