event ordering
Implementing this proved to be a huge pain in the ass because I forgot
that the values were stored in a dictionary at one point and
dictionaries were unordered. I found all sorts of other problems and
possible causes before I finally got to that part of the code...

Did require a database schema change and the passing of both lists and
dicts to complete, but not a huge deal except for the trouble it (should
not have) taken to implement.