implement searching for events
The code is a bit messy perhaps; it does need to be considerably
different than that for searching for entries, but it's not out of the
question that in a future refactor they could be edited to use some of
the same code.

It does work, however, and is accessible in a sensible way.